controllers: remove the storagecluster controller
It was created to watch the storagecluster and create the storagesystem
for that storagecluster.

As storagesystem is being removed we don't need this controller.
